Add files via upload
This is the final version of this project. To Be fair, i am really excited about this idea and i am already re-writing the code from scratch. The fact is that i have attended a course in university where we talked in depth about networks and the internet and i had a better idea of how to create the Network Scanning part of the program. Also since i am working on other projects too i now have a better grasp of using multiple threads in a program in java and i will use this new knowledge on the newer version. Furthermore when writing this code i did not have a plan, even if my idea seemed like it would work i went into this with a lot of trial and error, and now that i have a better idea of how things work i can write the code again with a better approach, with better optimization. The newer version is almost finished but i don't have the time unfortunately to complete it now, since the new semester of university has already begun. My vision for the newer version is to write the code in java with a client program, and after some bug-fixing and making sure that the program is complete i would like to make it into an android application since i believe it would be much more useful in the mobile market and i would also like to learn how to code for android too. Therefore i don't think this version will have a java version, at least not a GUI for desktops, it will only run on console.